sample updated

+++ database/2_bdd_intwetion.sql
@@ -19,7 +19,7 @@ CREATE TABLE `questions`
 (
   `id` int UNIQUE PRIMARY KEY AUTO_INCREMENT,
   `id_card` int,
-  `number_question` int,
+  `number_question` float,
   `text_question` text,
   `image_question` varchar(255),
   `type_response` int,
